recipes/data-integrity/model_comparison/requirements.txt
kaleido>=0.2.0   # For static image export from plotly (if write_image=True)
matplotlib>=3.4.0

nbformat>=4.2.0

# NLP and text processing
nltk>=3.6.0
numpy>=1.21.0

# Additional dependencies that may be needed
openpyxl>=3.0.0  # For Excel file support in pandas
# Core dependencies
pandas>=1.3.0
plotly>=5.0.0
rouge>=1.0.0
scikit-learn>=1.0.0
seaborn>=0.11.0
sentence-transformers>=2.2.0
spacy>=3.0.0
textblob>=0.17.0
textstat>=0.7.0

# Optional but recommended for full functionality
umap-learn>=0.5.0
